% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitative_models.R
\name{adjoint}
\alias{adjoint}
\title{Adjoint matrix}
\usage{
adjoint(A)
}
\arguments{
\item{A}{A square matrix}
}
\value{
The adjoint matrix of A
}
\description{
Adjoint matrix
}
\details{
Compute the adjoint matrix
}
\examples{
n <- 13
A <- matrix(c(-1, 0, 0, 1, 1, 0, 0, -1, -1, -1, 0, 0, -1, 0, -1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, -1, 0, 0, 0, 1, -1, 0, 0, 0, 0, 0, -1, 1, 0, -1, 1, 1, 0, 0, 0, 0, -1, 0, 0, 0, 0, 0, 0, -1, -1, 0, 0, 0, 0, 0, 1, 0, 0, -1, 0, 0, 1, -1, 0, 0, 0, 0, 0, 0, 0, 0, 0, -1, 0, 1, 1, -1, -1, -1, -1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, -1, 1, 0, 0, 0, 0,1, 0, 0, 0, 0, 0, 1, -1, -1, 0, 1, 0, 0,0, 0, 0, 0, 0, 0, 0, 0, 0, -1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, -1, 0, -1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, -1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, -1),n ,n ,byrow=TRUE)
adjoint(A)
}
