% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitative_models.R
\name{Tmat}
\alias{Tmat}
\title{Absoulte feedback (T)}
\usage{
Tmat(A)
}
\arguments{
\item{A}{A square matrix}
}
\value{
The absolute feedback of a matrix A
}
\description{
Absoulte feedback (T)
}
\details{
Compute the absoulte feedback of a matrix
}
\examples{
n <- 13
A <- matrix(c(-1, 0, 0, 1, 1, 0, 0, -1, -1, -1, 0, 0, -1, 0, -1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, -1, 0, 0, 0, 1, -1, 0, 0, 0, 0, 0, -1, 1, 0, -1, 1, 1, 0, 0, 0, 0, -1, 0, 0, 0, 0, 0, 0, -1, -1, 0, 0, 0, 0, 0, 1, 0, 0, -1, 0, 0, 1, -1, 0, 0, 0, 0, 0, 0, 0, 0, 0, -1, 0, 1, 1, -1, -1, -1, -1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, -1, 1, 0, 0, 0, 0,1, 0, 0, 0, 0, 0, 1, -1, -1, 0, 1, 0, 0,0, 0, 0, 0, 0, 0, 0, 0, 0, -1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, -1, 0, -1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, -1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, -1),n ,n ,byrow=TRUE)
Tmat(A)
}
